+++ client/src/app/auth/pages/transactions/transactions.tsx
@@ -6,8 +6,14 @@ import { useQuery } from '@tanstack/react-query';
 import { AxiosResponse } from 'axios';
 import { Transaction } from '@/types/transaction';
 import TransactionCards from './transaction-cards';
+import { SortDirection } from './transactions-header/sort-button';
+import TransactionsHeader from './transactions-header/transactions.header';
+import { Sorts } from './transactions-header/sort-by-menu';
 
 const Transactions = (): JSX.Element => {
+  const [sort, setSort] = React.useState(Sorts.Date);
+  const [sortDirection, setSortDirection] = React.useState(SortDirection.Decending);
+
   const { request } = React.useContext<any>(AuthContext);
 
   const transactionsQuery = useQuery({
@@ -33,7 +39,18 @@ const Transactions = (): JSX.Element => {
   return (
     <div className="flex w-full flex-col items-center gap-2">
       <EmailVerified />
-      <TransactionCards transactions={transactionsQuery.data ?? []} />
+      <TransactionsHeader
+        transactions={transactionsQuery.data ?? []}
+        sort={sort}
+        setSort={setSort}
+        sortDirection={sortDirection}
+        setSortDirection={setSortDirection}
+      />
+      <TransactionCards
+        transactions={transactionsQuery.data ?? []}
+        sort={sort}
+        sortDirection={sortDirection}
+      />
     </div>
   );
 };
